Branch edit

+++ src/main/java/seedu/address/logic/commands/DeleteStallReviewCommand.java
@@ -21,11 +21,12 @@ public class DeleteStallReviewCommand extends Command {
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
             + ": Deletes the stall review identified by the index number used in the displayed stall list.\n"
-            + PREFIX_STALL + "STALL NUMBER "
+            + PREFIX_STALL + "STALL_INDEX "
             + "Example: " + COMMAND_WORD + ""
             + PREFIX_STALL + "1";
 
     public static final String MESSAGE_DELETE_STALL_REVIEW_SUCCESS = "Deleted Stall Review: %1$s";
+    private static final String MESSAGE_ITEM_REVIEW_NOT_FOUND = "This stall does not have a review.";
 
     private final Index targetIndex;
 
@@ -43,6 +44,11 @@ public class DeleteStallReviewCommand extends Command {
         }
 
         Stall stallToDeleteReview = lastShownList.get(targetIndex.getZeroBased());
+
+        if (!stallToDeleteReview.hasStallReview()) {
+            throw new CommandException(MESSAGE_ITEM_REVIEW_NOT_FOUND);
+        }
+
         stallToDeleteReview.deleteReview();
         return new CommandResult(String.format(MESSAGE_DELETE_STALL_REVIEW_SUCCESS,
                 Messages.format(stallToDeleteReview)));
